fix(api): detect an in-use port by the Winsock errno on Windows
The #69 port walk never walked on Windows. isAddrInUse compared against
syscall.EADDRINUSE, but Winsock reports the condition as WSAEADDRINUSE, and on
Windows syscall.EADDRINUSE is a synthesised APPLICATION_ERROR with nothing
mapping the Winsock errno back to it — so the comparison was always false and
the daemon exited instead of trying the next port.

Split the predicate per platform: the !windows build keeps the syscall.EADDRINUSE
check, and the Windows build compares golang.org/x/sys/windows.WSAEADDRINUSE.

CI never caught it because ubuntu runners do not compile internal/api's
_windows.go half at all, so a break there surfaced at release time rather than
in CI. A Backend (Windows) job now builds, vets and tests internal/api on
windows-latest, which both links and exercises the Winsock path. internal/api
needs no cgo — only internal/api/capture's pcap path does — so it runs with
CGO_ENABLED=0.

Closes #1537
