+++ internal/api/server_port_fallback_unix.go
@@ -1,0 +1,14 @@
+//go:build !windows
+
+package api
+
+import (
+	"errors"
+	"syscall"
+)
+
+// isAddrInUse reports whether a bind failed because something else already
+// holds the port.
+func isAddrInUse(err error) bool {
+	return errors.Is(err, syscall.EADDRINUSE)
+}
