+++ internal/api/server_port_fallback_windows.go
@@ -1,0 +1,18 @@
+package api
+
+import (
+	"errors"
+
+	"golang.org/x/sys/windows"
+)
+
+// isAddrInUse reports whether a bind failed because something else already
+// holds the port.
+//
+// Winsock reports that as WSAEADDRINUSE, which is not the value
+// syscall.EADDRINUSE carries here: on Windows that constant is a synthesised
+// APPLICATION_ERROR with nothing mapping the Winsock errno back to it, so
+// comparing against it is always false and the walk would never happen (#1537).
+func isAddrInUse(err error) bool {
+	return errors.Is(err, windows.WSAEADDRINUSE)
+}
